Optimizing nanotime impl a bit

+++ timer/mono.go
@@ -17,7 +17,7 @@ package timer
 // Now returns the current monotonic time, with an arbitrary starting time.
 // It is not useful for telling the current wall clock time
 func Now() uint64 {
-	return uint64(nanotime())
+	return nanotime()
 }
 
 // Since returns the duration in nanoseconds since the given starting time.
