+++ timer/mono_fallback.go
@@ -25,6 +25,6 @@ func monotime() (sec int64, nsec int32) {
 	return now.Unix(), int32(now.Nanosecond())
 }
 
-func nanotime() int64 {
-	return time.Now().UnixNano()
+func nanotime() uint64 {
+	return uint64(time.Now().UnixNano())
 }
